Refactors subscriber to package by feature
