+++ coinbase-integration/gradle.properties
@@ -1,6 +1,6 @@
 kotlinVersion=1.4.10
 shadowJarPluginVersion=6.0.0
-http4kVersion=3.283.0
+http4kVersion=4.4.0.1
 http4kStarterVersion=1.2
 typesafeVersion=1.4.0
 slf4jVersion=2.0.0-alpha1
