--- coinbase-integration/subscriber/src/main/kotlin/me/pysquad/cryptobot/subscriber/AppConfiguration.kt
@@ -1,11 +1,0 @@
-package me.pysquad.cryptobot.subscriber
-
-import com.typesafe.config.Config
-import com.typesafe.config.ConfigFactory
-
-val configReader: Config = ConfigFactory.load()
-
-data class AppConfiguration(
-        val port: Int = configReader.getInt("application.port"),
-        val host: String = configReader.getString("application.host")
-)
