+++ coinbase-integration/subscriber/src/main/kotlin/me/pysquad/cryptobot/subscriber/Application.kt
@@ -1,13 +1,6 @@
 package me.pysquad.cryptobot.subscriber
 
 import me.pysquad.cryptobot.http4k.starter.http4kApp
-import me.pysquad.cryptobot.subscriber.coinbase.CoinbaseApi
-import me.pysquad.cryptobot.subscriber.coinbase.CoinbaseConfiguration
-import me.pysquad.cryptobot.subscriber.endpoints.subscriberRoutes
-import me.pysquad.cryptobot.subscriber.repo.SubscriberRepo
-import me.pysquad.cryptobot.subscriber.rethinkdb.RethinkDbConfiguration
-import me.pysquad.cryptobot.subscriber.rethinkdb.RethinkDbDatasource
-import me.pysquad.cryptobot.subscriber.service.SubscriberService
 
 fun main() {
     val appConfig = AppConfiguration()
@@ -17,13 +10,13 @@ fun main() {
         val coinbaseConfig = CoinbaseConfiguration()
         val rethinkDbConfig = RethinkDbConfiguration()
 
-        val service = SubscriberService(
+        val subscriberService = SubscriberService(
                 CoinbaseApi.client(coinbaseConfig),
                 coinbaseConfig,
-                SubscriberRepo.impl(RethinkDbDatasource(rethinkDbConfig))
+                SubscriberRepository.impl(RethinkDbDatasource(rethinkDbConfig))
         )
 
         // passing the routes
-        subscriberRoutes(service)
+        SubscriberWebController(subscriberService).routes()
     }.run()
 }
