+++ coinbase-integration/subscriber/src/main/kotlin/me/pysquad/cryptobot/subscriber/SubscriberService.kt
@@ -1,24 +1,16 @@
-package me.pysquad.cryptobot.subscriber.service
+package me.pysquad.cryptobot.subscriber
 
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import me.pysquad.cryptobot.subscriber.coinbase.CoinbaseApi
-import me.pysquad.cryptobot.subscriber.coinbase.CoinbaseConfiguration
-import me.pysquad.cryptobot.subscriber.coinbase.model.CoinbaseWsResponse
-import me.pysquad.cryptobot.subscriber.coinbase.model.CoinbaseWsSubscribeRequest
-import me.pysquad.cryptobot.subscriber.endpoints.model.SubscriberResponse
-import me.pysquad.cryptobot.subscriber.model.CoinbaseMessage
-import me.pysquad.cryptobot.subscriber.model.ProductIds
-import me.pysquad.cryptobot.subscriber.repo.SubscriberRepo
 import org.http4k.format.Jackson.auto
 import org.http4k.websocket.WsMessage
 import org.slf4j.LoggerFactory
 import kotlin.concurrent.thread
 
 class SubscriberService(
-        private val coinbaseApi: CoinbaseApi,
-        private val coinbaseConfig: CoinbaseConfiguration,
-        private val subscriberRepo: SubscriberRepo
+    private val coinbaseApi: CoinbaseApi,
+    private val coinbaseConfig: CoinbaseConfiguration,
+    private val subscriberRepository: SubscriberRepository
 ) {
 
     private val objectMapper = jacksonObjectMapper().registerModule(KotlinModule())
@@ -41,7 +33,7 @@ class SubscriberService(
 
             return if (wsFeedResp.type != "error") {
                 // right now we only send one channel, which is a 'ticker' channel
-                subscriberRepo.storeSubscriptions(wsFeedResp.channels.first().productIds)
+                subscriberRepository.storeSubscriptions(wsFeedResp.channels.first().productIds)
                 thread { received().storeInChunks() }
 
                 SubscriberResponse.success(
@@ -68,7 +60,7 @@ class SubscriberService(
 
         for(listOfMessages in chunked(6)) {
             val coinbaseMessages = listOfMessages.map(wsLens::extract)
-            subscriberRepo.storeMessages(coinbaseMessages)
+            subscriberRepository.storeMessages(coinbaseMessages)
         }
     }
 }
