--- coinbase-integration/subscriber/src/main/kotlin/me/pysquad/cryptobot/subscriber/endpoints/SubscriberRoutes.kt
@@ -1,7 +1,0 @@
-package me.pysquad.cryptobot.subscriber.endpoints
-
-import me.pysquad.cryptobot.subscriber.service.SubscriberService
-
-fun subscriberRoutes(service: SubscriberService) = mutableListOf(
-        subscribe(service)
-)
